Make new Websocket message type.
The go type is WebSocketMessage, which has an event type and a message field, with JSON tags.

Use the message type in (*WebUI).WSBlockUpdater, which sends to the client.

Add messagesocket.js, which defines MessageSocket, the JS counter part  to the go type WebSocketMessage.
