+++ views/root.tmpl
@@ -26,6 +26,7 @@
     <link href="css/main.css" rel="stylesheet">
     <link rel="stylesheet/less" type="text/css" href="css/dynamic.less">
     <script src="js/Chart.min.js"></script>
+    <script src="js/messagesocket.js"></script>
     <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
     <!--[if lt IE 9]>
       <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
@@ -107,30 +108,39 @@
             };
         }
 	
-	var loc = window.location;
+	    var loc = window.location;
         var uri = 'ws:';
-
         if (loc.protocol === 'https:') {
             uri = 'wss:';
         }
         uri += '//' + loc.host;
         uri += loc.pathname + 'ws';
 
-        ws = new WebSocket(uri)
+        var ws = new MessageSocket(uri);
 
-        ws.onopen = function () {
+        ws.registerEvtHandler("open", function() {
             console.log('Connected')
             updateConnectionStatus('Connected', true);
-        }
-	
-	ws.onerror = function(evt) {
-            console.log("WebSocket errord.");
-            updateConnectionStatus('Disconnected', false);
-        };
+        });
 
-        ws.onmessage = function (evt) {
-            var m = JSON.parse(evt.data);
-            console.log("Received message:", m.height);
+        ws.registerEvtHandler("close", function() {
+            console.log('Disconnected')
+            updateConnectionStatus('Disconnected', false);
+        });
+
+        ws.registerEvtHandler("error", function(evt) {
+            console.log("WebSocket error:", evt);
+            updateConnectionStatus('Disconnected', false);
+        });
+
+        ws.registerEvtHandler("ping", function(evt) {
+            console.debug("ponging a ping")
+            ws.send("pong", "copy")
+        });
+
+        var updateMempool = function (event) {
+            console.log("Received mempoolsstxfeeinfo message", event);
+            var m = JSON.parse(event);
 	    
             // mempool fee info
             var number = document.getElementById('mempoolfeeinfo_number');
@@ -150,24 +160,33 @@
 
             var lm = document.getElementById('mempoolfeeinfo_lowestmineable');
             lm.innerHTML = String(m.lowest_mineable);
+        };
+
+        ws.registerEvtHandler("mempoolsstxfeeinfo", updateMempool);
+
+        var updateBlockData = function (event) {
+            console.log("Received newblock message", event);
+            var newblock = JSON.parse(event);
 
             // block summary data
+            var b = newblock.block;
+
             var height = document.getElementById('blockheight');
-            height.innerHTML = String(m.height);
+            height.innerHTML = String(b.height);
 
             var hash = document.getElementById('blockhash');
-            hash.innerHTML = m.hash;
+            hash.innerHTML = b.hash;
 
             var blockhash = document.getElementById('blocksize');
-            blockhash.innerHTML = String(m.size);
+            blockhash.innerHTML = String(b.size);
 
             var diff = document.getElementById('blockdiff');
-            diff.innerHTML = String(m.diff);
+            diff.innerHTML = String(b.diff);
 
             var sdiff = document.getElementById('blocksdiff');
-            sdiff.innerHTML = String(m.sdiff);
+            sdiff.innerHTML = String(b.sdiff);
 	    
-	    var poolsize = document.getElementById('poolsize');
+	        var poolsize = document.getElementById('poolsize');
             poolsize.innerHTML = String(b.ticket_pool.size);
 
             var poolvalue = document.getElementById('poolvalue');
@@ -176,8 +195,9 @@
             var poolvalavg = document.getElementById('poolvalavg');
             poolvalavg.innerHTML = String(b.ticket_pool.valavg);
 
-            // stake data
-            var s = m.stake;
+            // stake
+            var s = newblock.stake;
+
             var sdiffcurrent = document.getElementById('blocksdiff');
             sdiffcurrent.innerHTML = String(s.stakediff.current);
             // var sdiffnext = document.getElementById('blocksdiffnext');
@@ -205,13 +225,15 @@
             feeinfomedian.innerHTML = String(feeinfo.median);
             var feeinfomean = document.getElementById('feeinfo_mean');
             feeinfomean.innerHTML = String(feeinfo.mean);
-        }
+        };
 
-        // setInterval(function () {
-        //     ws.send('Hello, Server!');
-        // }, 1000);
+        ws.registerEvtHandler("newblock", updateBlockData);
+
+        setInterval(function () {
+            ws.send("ping", 'Hi. I am a client!');
+        }, 1000);
 	
-	function updateConnectionStatus(msg, connected) {
+	    function updateConnectionStatus(msg, connected) {
             var el = document.querySelector('#connection');
             if (connected) {
                 if (el.classList) {
@@ -233,10 +255,10 @@
             el.innerHTML = msg + '<div></div>';
         }
 
-        function closeConnection() {
-            ws.close();
-            updateConnectionStatus('Disconnected', false);
-        }
+        // function closeConnection() {
+        //     ws.close();
+        //     updateConnectionStatus('Disconnected', false);
+        // }
     </script>
 
     <!-- Fixed navbar -->
@@ -263,7 +285,7 @@
                     <li><a href="https://github.com/dcrdata/dcrdata#json-rest-api" title="API Endpoints" title="_blank">API</a></li>
                     <li><a href="https://api.dcrdata.net/status" title="API Status" title="_blank">API Status</a></li>
                     <li><a href="https://decred.slack.com/" title="Decred Slack" title="_blank">Decred Slack</a></li>
-                    <li><span id="connection" onclick="closeConnection()">Connecting to event source...<div></div></span></li> 
+                    <li><span id="connection">Connecting to event source...<div></div></span></li> 
                 </ul>
             </div>
             <!--/.nav-collapse -->
